# Adding optimized manual implementation of WriteByte