--- a/src/RecyclableMemoryStream.cs
+++ b/src/RecyclableMemoryStream.cs
@@ -68,12 +68,6 @@ namespace Microsoft.IO
         /// </summary>
         private readonly List<byte[]> blocks = new List<byte[]>(1);
 
-        /// <summary>
-        /// This buffer exists so that WriteByte can forward all of its calls to Write
-        /// without creating a new byte[] buffer on every call.
-        /// </summary>
-        private readonly byte[] byteBuffer = new byte[1];
-
         private readonly Guid id;
 
         private readonly RecyclableMemoryStreamManager memoryManager;
@@ -626,10 +620,33 @@ namespace Microsoft.IO
         /// <exception cref="ObjectDisposedException">Object has been disposed</exception>
         public override void WriteByte(byte value)
         {
-            this.CheckDisposed();
-            this.byteBuffer[0] = value;
-            this.Write(this.byteBuffer, 0, 1);
-        }
+			this.CheckDisposed();
+			int blockSize = this.memoryManager.BlockSize;
+			int end = this.position + 1;
+			// Check for overflow
+			if (end > MaxStreamLength)
+			{
+				throw new IOException("Maximum capacity exceeded");
+			}
+
+			this.EnsureCapacity(end);
+
+			if (this.largeBuffer is null)
+			{
+				int block = this.position / blockSize, offset = this.position % blockSize;
+
+				byte[] currentBlock = this.blocks[block];
+
+				currentBlock[offset] = value;
+			}
+			else
+			{
+				this.largeBuffer[this.position] = value;
+			}
+
+			this.position = end;
+			this.length = Math.Max(this.position, this.length);
+		}
 
         /// <summary>
         /// Reads a single byte from the current position in the stream.