@clarkis117 2018-02-09
@doubleyewdee I poked, who I'm guessing is the maintainer on twitter a few weeks ago, this is the response I received: https://twitter.com/doubleyewdee/status/951285541319843840
Although, I should probably update the benchmark to include results for .NET Core 1.x, 2.0, and 2.1; now that I got Benchmark.NET working with them on my dev box 👍