@benmwatson 2019-05-15
I am finally getting around to reviewing this... I'm curious what the motivation for this is. When I originally wrote this, I think I assumed that WriteByte would be a rare call (because who writes to a stream one byte at a time?), thus preferred avoiding code duplication.

That said, I'm not opposed to this change.