@raymondpwilson 2019-06-13
I just discovered this branch and PR after independently optimising the WriteByte() method (in addition to a collection of other optimisations).

I have included my version of this function below if you are interested.

    public override void WriteByte(byte value)
    {
      this.CheckDisposed(); //<-- Editorial: The run time for this dominates WriteByte() performance. IFDEFing out this call (assuming your environment is safe to do so) markedly increases performance of this method.

      long end = (long)this.position + 1;

      // Check for overflow
      if (end > MaxStreamLength)
      {
        throw new IOException("Maximum capacity exceeded");
      }

      if (end >= this.capacity)  // <-- Note: Another optimisation that uses a cached capacity via backing variable rather than using the expensive Capacity property.
      {
        this.EnsureCapacity((int) end); // <- EnsureCapacity() is expensive - only call it if you need to
      }

      if (this.largeBuffer == null)
      {
        this.blocks[this.position / this.blockSize][this.position % this.blockSize] = value;
      }
      else
      {
        this.largeBuffer[this.position] = value;
      }

      this.position = (int)end;
      this.length = Math.Max(this.position, this.length);
    }