@benmwatson 2020-03-18
I'm not convinced of proper use cases for this. Writing 1MB stream one byte at a time feels like an artificial benchmark. I would prefer cleaner code right now.